+++ js/get-random-offer.js
@@ -7,55 +7,16 @@ import {
   getRandomLocation,
   getRandomLocationToString } from './util.js'
 
+import {
+  TITLES,
+  TYPES_OF_HOUSE,
+  CHECK_TIME,
+  FEATURES,
+  DESCRIPTIONS,
+  PHOTOS
+} from './mock.js'
+
 const OFFERS_COUNT = 10
-
-const TITLES = [
-  'Милая, уютная квартирка в Токио',
-  'Отличное предложение по отличной цене',
-  'Вы никогда не встречали такую квартиру',
-  'Центр Токио только для Вас',
-  'Жильё в мегаполисе',
-];
-
-const TYPES_OF_HOUSE =
-[
-  'palace',
-  'flat',
-  'house',
-  'bungalow',
-];
-
-const CHECK_TIME =
-[
-  '12:00',
-  '13:00',
-  '14:00',
-]
-
-const  FEATURES =
-[
-  'wifi',
-  'dishwasher',
-  'parking',
-  'washer',
-  'elevator',
-  'conditioner',
-]
-
-const DESCRIPTIONS =
-[
-  'Милая квартирка в Токио, о которой вы мечтали всю жизнь, уже ждёт вас!',
-  'За небольшую цену вы получаете максимум удобств и шикарный вид из окна',
-  'Вы никогда не встречали такую квартиру, позвоните и узайте подробнее!',
-]
-
-const PHOTOS =
-[
-  'http://o0.github.io/assets/images/tokyo/hotel1.jpg',
-  'http://o0.github.io/assets/images/tokyo/hotel2.jpg',
-  'http://o0.github.io/assets/images/tokyo/hotel3.jpg',
-]
-
 
 let getRandomOffer = () => {
   return   {
